Use shared integration setup helpers
* Move namespace clean up to utils
* Use GinkgoDefer for integration manager start
  copying approach from controller-runtime tests

[#173218516](https://www.pivotaltracker.com/story/show/173218516)
