+++ integration/environment/environment.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	gomegaConfig "github.com/onsi/ginkgo/config"
+	"github.com/onsi/gomega"
 	"github.com/pkg/errors"
 	"github.com/spf13/afero"
 
@@ -48,7 +49,7 @@ func NewEnvironment(kubeConfig *rest.Config) *Environment {
 	// the single namespace used by this test
 	ns := utils.GetNamespaceName(namespaceID)
 
-	return &Environment{
+	env := &Environment{
 		Environment: &utils.Environment{
 			ID:         namespaceID,
 			Namespace:  ns,
@@ -66,6 +67,10 @@ func NewEnvironment(kubeConfig *rest.Config) *Environment {
 			Machine: machine.NewMachine(),
 		},
 	}
+	gomega.SetDefaultEventuallyTimeout(env.PollTimeout)
+	gomega.SetDefaultEventuallyPollingInterval(env.PollInterval)
+
+	return env
 }
 
 // SetupClientsets initializes kube clientsets
